+++ tests/test_cli.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 import pytest
 
 import mdformat
-from mdformat._cli import get_package_name, get_plugin_versions, run, wrap_paragraphs
+from mdformat._cli import get_package_name, get_plugin_info_str, run, wrap_paragraphs
 from mdformat.plugins import CODEFORMATTERS
 
 UNFORMATTED_MARKDOWN = "\n\n# A header\n\n"
@@ -350,16 +350,20 @@ def test_get_package_name():
     assert get_package_name(mdformat) == "mdformat"
 
 
-def test_get_plugin_versions():
-    # Pretend that "pytest" and "unittest.mock.patch" are plugins
-    versions = get_plugin_versions({"p1": pytest}, {"f1": patch})  # type: ignore[dict-item] # noqa: E501
-    assert versions[0][0] == "pytest"
-    assert versions[0][1] != "unknown"
-    assert versions[1] == ("unittest", "unknown")
+def test_get_plugin_info_str():
+    info = get_plugin_info_str(
+        {"mdformat-tables": ("0.1.0", ["tables"])},
+        {"mdformat-black": ("12.1.0", ["python"])},
+    )
+    assert (
+        info
+        == """\
+installed codeformatters:
+  mdformat-black: python
 
-    with patch("mdformat._cli.inspect.getmodule", return_value=None):
-        versions = get_plugin_versions({"p1": pytest}, {})  # type: ignore[dict-item]
-    assert versions[0] == ("unknown", "unknown")
+installed extensions:
+  mdformat-tables: tables"""
+    )
 
 
 def test_no_timestamp_modify(tmp_path):